backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
ipykernel==6.30.1
json5==0.12.1
motor==3.3.1
python-dotenv>=1.0.1
pydantic>=2.6.4
pymongo==4.5.0
requests-2.32.5